combined Profile & Account into a single page
1. Removed Profile button from user menu.
2.  Renamed Account as Account Settings in i18n.json .
3. Removed Profile button from menu.
4. Added Username, Registered e-mail and Company fields and their update functionality under Account Settings.
5.  Moved profile helper functions from profile.js to account.js
6. moved autoform.js file from profile to account.
7. Deleted profile.html
8. Deleted profile.js
9. Deleted profile.less
10. Deleted profile folder.
