+++ apinf_packages/core/client/userMenu.html
@@ -20,11 +20,6 @@ https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence
     </a>
     <ul class="dropdown-menu">
       <li>
-        <a href="{{ pathFor 'profile' }}" id="profile-button">
-          {{_ "userMenu_profile" }}
-        </a>
-      </li>
-      <li>
         <a href="{{ pathFor 'account' }}" id="account-button">
           {{_ "userMenu_account" }}
         </a>
