+++ apinf_packages/core/lib/i18n/en.i18n.json
@@ -19,7 +19,7 @@
   "account_delete_warning": "There's no coming back from this one",
   "account_delete_title": "Delete account",
   "account_details": "Account Details",
-  "accountPage_title_account": "Account",
+  "accountPage_title_account": "Account Settings",
   "addApiPage_title_addApi": "Add API",
   "addApi_button_text": "Add API",
   "addApi_header": "Add an API",
@@ -1171,7 +1171,7 @@
   "settings_errorMessage": "Invalid Settings! Something is not right or a field is missing.",
   "settingsPage_title_settings": "Settings",
   "settingsMenu_profile": "Profile",
-  "settingsMenu_account": "Account",
+  "settingsMenu_account": "Account Settings",
   "settingsMenu_branding": "Branding",
   "settingsMenu_settings": "Settings",
   "settingsMenu_proxies": "Proxies",
@@ -1226,7 +1226,7 @@
   "uploadProjectLogo_acceptedExtensions": "Only .jpg, .jpeg, .png and .gif file formats are allowed.",
   "uploadProjectLogo_successfully_uploaded": "Your project logo successfully uploaded!",
   "uploadProjectLogo_successfully_deleted": "Your project logo successfully deleted!",
-  "userMenu_account": "Account",
+  "userMenu_account": "Account Settings",
   "userMenu_branding": "Branding",
   "userMenu_profile": "Profile",
   "userMenu_proxies": "Proxies",
