+++ apinf_packages/settings/client/menu/menu.html
@@ -7,13 +7,6 @@ https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence
   <div id="settings-menu" class="col-md-3">
     <ul class="list-group">
       <a
-        href="{{ pathFor route='profile' }}"
-        class="list-group-item {{ isActiveRoute 'profile' }}"
-        id="button-profile">
-          {{_ "settingsMenu_profile" }}
-          <span class="glyphicon glyphicon-chevron-right pull-right"></span>
-      </a>
-      <a
         href="{{ pathFor route='account' }}"
         class="list-group-item {{ isActiveRoute 'account' }}"
         id="button-account">
