--- apinf_packages/users/client/profile/profile.html
@@ -1,42 +1,0 @@
-<!-- Copyright 2017 Apinf Oy
-This file is covered by the EUPL license.
-You may obtain a copy of the licence at
-https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 -->
-
-<template name="profile">
-  <div class="container padded-container">
-    <div class="row">
-      {{> settingsMenu }}
-
-      <div class="col-md-6">
-        <div class="panel panel-default">
-          <div class="panel-heading">
-            <div class="panel-title">
-              {{_ "profile_Header" }}
-            </div>
-          </div>
-          <div class="panel-body">
-            {{# autoForm
-              id="updateProfile"
-              collection=usersCollection
-              doc=currentUser
-              type="update" }}
-              {{> afQuickField name="username" id="username" }}
-
-              <div class="form-group">
-                <label for="email">
-                  {{_ "profile_labelText_registeredEmail" }}
-                </label>
-                <p>{{ userEmail }}</p>
-              </div>
-              {{> afQuickField name="profile.company" id="company" }}
-              <button type="submit" class="btn btn-success profile-update" id="update-button">
-                {{_ "profile_UpdateButton" }}
-              </button>
-            {{/ autoForm }}
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
