--- apinf_packages/users/client/profile/profile.js
@@ -1,54 +1,0 @@
-/* Copyright 2017 Apinf Oy
-This file is covered by the EUPL license.
-You may obtain a copy of the licence at
-https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 */
-
-// Meteor packages imports
-import { Meteor } from 'meteor/meteor';
-import { Template } from 'meteor/templating';
-
-// Meteor contributed packages imports
-import { DocHead } from 'meteor/kadira:dochead';
-import { TAPi18n } from 'meteor/tap:i18n';
-
-// Collection imports
-import Branding from '/apinf_packages/branding/collection';
-
-Template.profile.onCreated(function () {
-  // Get reference to template instance
-  const instance = this;
-
-  instance.autorun(() => {
-    // Get Branding collection content
-    const branding = Branding.findOne();
-    // Check if Branding collection and siteTitle are available
-    if (branding && branding.siteTitle) {
-      // Set the page title
-      const pageTitle = TAPi18n.__('profilePage_title_profile');
-      DocHead.setTitle(`${branding.siteTitle} - ${pageTitle}`);
-    }
-  });
-});
-
-Template.profile.helpers({
-  currentUser () {
-    return Meteor.user();
-  },
-  usersCollection () {
-    // Return reference to Meteor.users collection
-    return Meteor.users;
-  },
-  userEmail () {
-    let email;
-    // Get current user
-    const user = Meteor.user();
-
-    // Make sure user exists
-    if (user) {
-      // Get e-mail address
-      email = user.emails[0].address;
-    }
-
-    return email;
-  },
-});
