--- apinf_packages/users/client/profile/profile.less
@@ -1,14 +1,0 @@
-/* Copyright 2017 Apinf Oy
-This file is covered by the EUPL license.
-You may obtain a copy of the licence at
-https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 */
-
-#copyApi{
-  margin-top: 0.62em;
-}
-
-input[name='profile.apiKey']{
-  width: 100%;
-  border: 0px solid #fff;
-  outline: none;
-}
